requirements.txt

Django~=5.1.7
djangorestframework~=3.15.2
pytest~=8.3.5
django-fsm~=3.0.0
pytest